+++ anndata/h5py/__init__.py
@@ -1,9 +1,9 @@
 """Wraps `h5py <http://www.h5py.org/>`_ to handle sparse matrices.
 
-:mod:`anndata.h5py` is based on and uses the conventions of `h5sparse
-<https://github.com/appier/h5sparse>`_ by `Appier
-Inc. <https://www.appier.com/>`_. See the copyright and license note in the
-source code.
+:mod:`anndata.h5py` is based on and uses the conventions of
+`h5sparse <https://github.com/appier/h5sparse>`_ by
+`Appier Inc. <https://www.appier.com/>`_.
+See the copyright and license note in the source code.
 
 The design choices of :mod:`anndata.h5py`, however, are different. In
 particular, :mod:`anndata.h5py` allows handling sparse and non-sparse data at
@@ -26,4 +26,4 @@ from h5py import Dataset, special_dtype
 
 # Problem: the H5py intersphinx is broken, and only contains e.g. `Dataset` directly.
 # So we can’t possibly link to it using :class:`Dataset`, since that will always find our version.
-Dataset.__doc__ = """Equivalent to :ref:`h5py.Dataset <h5py:Dataset>`."""
+Dataset.__doc__ = """Equivalent to :class:`h5py.Dataset <h5py:Dataset>`."""