+++ anndata/h5py/h5sparse.py
@@ -1,7 +1,7 @@
 # TODO:
 # - think about making all of the below subclasses
 # - think about supporting the COO format
-from typing import Optional
+from typing import Optional, Union, KeysView
 
 import six
 import h5py
@@ -53,14 +53,14 @@ def _load_h5_dataset_as_sparse(sds, chunk_size=6000):
 
 
 class Group:
-    """Like :ref:`h5py.Group <h5py:Group>`, but able to handle sparse matrices.
+    """Like :class:`h5py.Group <h5py:Group>`, but able to handle sparse matrices.
     """
 
     def __init__(self, h5py_group, force_dense=False):
         self.h5py_group = h5py_group
         self.force_dense = force_dense
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: str) -> Union[h5py.Group, h5py.Dataset, 'SparseDataset']:
         h5py_item = self.h5py_group[key]
         if isinstance(h5py_item, h5py.Group):
             if 'h5sparse_format' in h5py_item.attrs:
@@ -76,10 +76,10 @@ class Group:
     def __delitem__(self, name):
         self.h5py_group.__delitem__(name)
 
-    def __setitem__(self, key, value):
+    def __setitem__(self, key: str, value: Union[h5py.Group, h5py.Dataset, 'SparseDataset']):
         self.h5py_group.__setitem__(key, value)
 
-    def keys(self):
+    def keys(self) -> KeysView[str]:
         return self.h5py_group.keys()
 
     def create_dataset(self, name, data=None, chunk_size=6000, **kwargs):
@@ -118,7 +118,7 @@ Group.create_dataset.__doc__ = h5py.Group.create_dataset.__doc__
 
 
 class File(Group):
-    """Like :ref:`h5py.File <h5py:File>`, but able to handle sparse matrices.
+    """Like :class:`h5py.File <h5py:File>`, but able to handle sparse matrices.
     """
 
     def __init__(
@@ -237,7 +237,7 @@ _cs_matrix._zero_many = _zero_many
 
 
 class SparseDataset(IndexMixin):
-    """Analogous to :ref:`h5py.Dataset <h5py:Dataset>`, but for sparse matrices.
+    """Analogous to :class:`h5py.Dataset <h5py:Dataset>`, but for sparse matrices.
     """
 
     def __init__(self, h5py_group):