+++ docs/release_notes.rst
@@ -7,7 +7,7 @@ See all releases `here <https://github.com/theislab/anndata/releases>`_. The fol
 2. initialatization from pandas DataFrames
 3. iteration over chunks :func:`~anndata.AnnData.chunked_X` and :func:`~anndata.AnnData.chunk_X`
 4. support for reading zarr files: :func:`~anndata.read_zarr`
-5. changed default compression to `None` in :func:`~anndata.AnnData.write_h5ad` to speed up read and write, disk space use is usually less critical (v0.6.16)
+5. changed default compression to ``None`` in :func:`~anndata.AnnData.write_h5ad` to speed up read and write, disk space use is usually less critical (v0.6.16)
       
 
 **May 1, 2018**: version 0.6
@@ -24,7 +24,7 @@ See all releases `here <https://github.com/theislab/anndata/releases>`_. The fol
 
 1. inform about duplicates in :class:`~anndata.AnnData.var_names` and resolve them using :func:`~anndata.AnnData.var_names_make_unique`
 2. automatically remove unused categories after slicing
-3. read/write `.loom` files using loompy 2
+3. read/write ``.loom`` files using loompy 2
 4. fixed read/write for a few text file formats
 5. read `UMI tools <https://github.com/CGATOxford/UMI-tools>`__ files: :func:`~anndata.read_umi_tools`